test(trade-ideas): split conflict-safety tests from paper reconciliation
test_paper_reconciliation.py (420 lines, allowlisted) bundled two concerns:
core matching/dedup/rerun behavior and the conflict-rejection / no-fallback
safety contract (unknown decision_id must not fall back to client_order_id
or symbol/side; symbol/side conflicts are rejected without mutation).

Extract the 6 safety tests to test_paper_reconciliation_conflicts.py. Both
files share only conftest fixtures (no local coupling), so the split is
clean: original 261 lines, new file 185, all 16 tests unchanged.

Allowlist: drop the now-stale reconciliation entry. Refine the remaining
test_strategy_engine_orders.py reason — its file-local helpers are shared
across every test group, so it is intentionally kept whole rather than
fragmented behind a shared helper module (not "split pending").
